refactor(domain): 결과 출력에 대한 메서드 추출
